+++ src/main/java/io/temporal/internal/common/HeaderUtils.java
@@ -1,0 +1,40 @@
+/*
+ *  Copyright (C) 2020 Temporal Technologies, Inc. All Rights Reserved.
+ *
+ *  Copyright 2012-2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ *  Modifications copyright (C) 2017 Uber Technologies, Inc.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"). You may not
+ *  use this file except in compliance with the License. A copy of the License is
+ *  located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ *  or in the "license" file accompanying this file. This file is distributed on
+ *  an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ *  express or implied. See the License for the specific language governing
+ *  permissions and limitations under the License.
+ */
+
+package io.temporal.internal.common;
+
+import io.temporal.proto.common.Header;
+import io.temporal.proto.common.Payload;
+import java.util.Map;
+
+public class HeaderUtils {
+
+  public static Header toHeaderGrpc(Map<String, Payload> headers) {
+    if (headers == null || headers.isEmpty()) {
+      return null;
+    }
+    Header.Builder builder = Header.newBuilder();
+    for (Map.Entry<String, Payload> item : headers.entrySet()) {
+      builder.putFields(item.getKey(), item.getValue());
+    }
+    return builder.build();
+  }
+
+  private HeaderUtils() {}
+}
